% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Binomal_Table_Density.R
\name{Binomial_Density}
\alias{Binomial_Density}
\title{Plot the density of binomial}
\usage{
Binomial_Density(n, frequency = TRUE, color = "#f8cea2",
  borderc = "#b78135")
}
\description{
Plot the density of binomial
}
