% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday_Problem.R
\name{Birthday_Problem}
\alias{Birthday_Problem}
\title{Birthday Problem}
\usage{
Birthday_Problem(size, days = 366)
}
\description{
Birthday Problem
}
\references{
\url{https://en.wikipedia.org/wiki/Birthday_problem}
}
