% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Kernel_Density.R
\name{Calculate_Kernel_Density}
\alias{Calculate_Kernel_Density}
\title{Calculate_Kernel_Density}
\usage{
Calculate_Kernel_Density(x, h = 1, x0 = seq(min(x), max(x), length =
  length(x)), kernel = c("rectangle", "triangle", "gauss"))
}
\description{
Calculate_Kernel_Density
}
