% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Binomial_Sample.R
\name{Create_Binomial_Table}
\alias{Create_Binomial_Table}
\title{Create a Binomial Frequency Table Based on the Number of Toss in a Binomial Distribution}
\usage{
Create_Binomial_Table(n)
}
\description{
Create a Binomial Frequency Table Based on the Number of Toss in a Binomial Distribution
}
