% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Html_Index.R
\name{Create_Index_Html}
\alias{Create_Index_Html}
\title{Create an Index HTML file
Create a HTML file listing the files in the current directory}
\usage{
Create_Index_Html(path = "~/Dropbox/Apps/updog/appendix")
}
\description{
Create an Index HTML file
Create a HTML file listing the files in the current directory
}
