% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Latin_Square_Matrix.R
\name{Create_Latin_Square_Matrix}
\alias{Create_Latin_Square_Matrix}
\title{Create a Latin Squared Matrix}
\usage{
Create_Latin_Square_Matrix(n = 5, Latin = TRUE)
}
\description{
Create a Latin Squared Matrix
}
