% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_Disjunction_Test_Stimuli_Generator.R
\name{Disjunction_Generate_Stimuli_List}
\alias{Disjunction_Generate_Stimuli_List}
\title{import magick 
import tuneR
Generate Test Stimuli List for the Disjunctions and the Modality Experiments}
\usage{
Disjunction_Generate_Stimuli_List(...)
}
\description{
import magick 
import tuneR
Generate Test Stimuli List for the Disjunctions and the Modality Experiments
}
