% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_Disjunction_Test_Stimuli_Generator.R
\name{Disjunction_Generate_Test_Audio}
\alias{Disjunction_Generate_Test_Audio}
\title{Generate Test Audios for the Disjunctions and the Modality Experiments}
\usage{
Disjunction_Generate_Test_Audio(row, ...)
}
\description{
Generate Test Audios for the Disjunctions and the Modality Experiments
}
