% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_Disjunction_Test_Stimuli_Generator.R
\name{Disjunction_Generate_Test_Image}
\alias{Disjunction_Generate_Test_Image}
\title{Generate Test Images for the Disjunctions and the Modality Experiments}
\usage{
Disjunction_Generate_Test_Image(row, ...)
}
\description{
Generate Test Images for the Disjunctions and the Modality Experiments
}
