% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Quantile_Quantile_Sample.R
\name{ECDF}
\alias{ECDF}
\title{Create the Empirical Cumulative Distribution Function (ECDF)}
\usage{
ECDF(data)
}
\arguments{
\item{data}{a numeric vector}
}
\description{
Create the Empirical Cumulative Distribution Function (ECDF)
}
\examples{
set.seed(1)
nrm100 <- rnorm(100, mean = 50, sd = 10)
ECDF(nrm100)
}
