% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Sound_Fragment.R
\name{Extract_Sound_Fragment}
\alias{Extract_Sound_Fragment}
\title{Extract Sound Fragment}
\usage{
Extract_Sound_Fragment(
  Original_File = "NaMe",
  Start_Offset = 0,
  Stop_Offset = 0,
  Sound_File_Path = "Objects_Audios",
  Play_New_File = TRUE
)
}
\description{
Extract Sound Fragment
}
