\name{IntAcMain}
\alias{IntAcMain}
\title{Draw Interaction and/or Main effects between two factors}
\usage{
IntAcMain(R = 2, C = 3, data = 1:6, adjust = 1)
}
\description{
Draw Interaction and/or Main effects between two factors.
}
\arguments{
  \item{R=2}{Levels of the row factor}
  \item{C=3}{Levels of the column factor}
  \item{data}{Data used to do the plotting}
  \item{adjust}{Zoom in or Zoom out the Y scale}
}
\examples{
IntAcMain()
}
