% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Julia_Call_Mixed_Models.R
\name{jglmer}
\alias{jglmer}
\title{Generalized Linear Mixed Effect Moddels}
\usage{
jglmer(formula, data, fast = TRUE, time = FALSE,
  family = "Binomial()", ...)
}
\description{
Generalized Linear Mixed Effect Moddels
}
