% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Julia_Call_Mixed_Models.R
\name{Julia_Mixed_Models}
\alias{Julia_Mixed_Models}
\title{Fit General Linear and Generalized Linear Mixed models from Julia}
\usage{
Julia_Mixed_Models(formula, data, fast = FALSE, time = TRUE,
  family = "Binomial()", model = "glmm")
}
\description{
Fit General Linear and Generalized Linear Mixed models from Julia
}
