% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Best_Language.R
\name{Language}
\alias{Language}
\title{Best Programming Language
What is the Best Programming Language in History?}
\usage{
Language()
}
\description{
Best Programming Language
What is the Best Programming Language in History?
}
