\name{Latin_Square_Design}
\alias{Latin_Square_Design}
\title{Show the Latin Square Experimental Design}
\description{
Show the Latin Square Experimental Design
}
\arguments{
  \item{n = 5}{Levels of the experimental factor}
  \item{Latin = TRUE}{Whether to use Latin Square method or not}
}
\usage{
Latin_Square_Design(n = 5, Latin = TRUE)
}
\examples{
Latin_Square_Design()
}
