\name{Modified_Frequency_Distribution_Mean}
\alias{Modified_Frequency_Distribution_Mean}
\title{Modified_Frequency_Distribution_Mean}
\description{
Modified_Frequency_Distribution_Mean
}
\arguments{
  \item{sample}{The sample to be plotted}
  \item{mu}{mu = 10}
  \item{sst}{sst = FALSE, to show the sample mean or not}
  \item{...}{Modified_Frequency_Distribution}
}
\usage{
Modified_Frequency_Distribution_Mean(sample, mu = 10, sst = FALSE, ...)
}
\examples{
Modified_Frequency_Distribution_Mean(sample, mu = 10, sst = FALSE, ...)
}
