% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation_Test.R
\name{Permutation_Test}
\alias{Permutation_Test}
\title{Compute the permutation test of correlation}
\usage{
Permutation_Test(p, q, size = 1000)
}
\arguments{
\item{p, q}{two data sets}

\item{size}{size = 1000, size of the permutation}
}
\description{
Compute the permutation test of correlation
}
\examples{
p <- c(5.0, 4.8, 4.7, 4.0, 5.3, 4.1, 5.5, 4.7, 3.3, 4.0, 4.0, 4.6, 5.3, 3.0, 3.5, 3.9, 4.7, 5.0, 5.2, 4.6)
q <- c(60, 59, 58, 47, 65, 48, 67, 70, 55, 63, 62, 65, 71, 56, 59, 60, 74, 77, 78, 62)
Permutation_Test(p, q)
}
