% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hausman_Test.R
\name{phtest_glmer}
\alias{phtest_glmer}
\title{Conduct a Hausman test}
\usage{
phtest_glmer(glmerMod, glmMod, ...)
}
\arguments{
\item{glmerMod}{Model}

\item{glmMod}{Model}
}
\description{
Conduct a Hausman test to see if a variable is treated as a fixed effect or a random effect
}
\references{
\url{https://stackoverflow.com/questions/23630214/hausmans-specification-test-for-glmer-from-lme4}
}
