% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Binomial_Density.R
\name{Plot_Binomial_Density}
\alias{Plot_Binomial_Density}
\title{Plot the density of a binomial distribution}
\usage{
Plot_Binomial_Density(xx = 0:10, size = 10, pi = c(0.2, 0.5, 0.8))
}
\description{
Plot the density of a binomial distribution
}
\examples{
Plot_Binomial_Density()
}
