% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Binomial_Sample.R
\name{Plot_Binomial_Sample}
\alias{Plot_Binomial_Sample}
\title{Plot The Sample Frequency of a Binomial Distribution}
\usage{
Plot_Binomial_Sample(number_of_toss, show_density = TRUE,
  show_smooth_curve = FALSE, ...)
}
\arguments{
\item{number_of_toss}{Number of toss}

\item{show_density}{To show the density}

\item{show_smooth_curve}{To show the normal smooth curve}
}
\description{
Plot The Sample Frequency of a Binomial Distribution
}
