% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Binomial_Simulate.R
\name{Plot_Binomial_Simulate}
\alias{Plot_Binomial_Simulate}
\title{Plot the data stimulating the binomial distribution}
\usage{
Plot_Binomial_Simulate(samples = 10000, size = 20, pi = 0.5)
}
\description{
Plot the data stimulating the binomial distribution
}
\examples{
Plot_Binomial_Simulate(pi = 0.3, size = 5)
}
