% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Boxplot.R
\name{Plot_Boxplot}
\alias{Plot_Boxplot}
\title{Plot a boxplot}
\usage{
Plot_Boxplot(data)
}
\arguments{
\item{data}{a numeric vector}
}
\description{
Illustrate how a boxplot and five number summary are calculated
}
\examples{
fat <- c(PASWR::Bodyfat[["fat"]], 60)
Plot_Boxplot(fat)
}
