% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compare_Binomial_Poisson.R
\name{Plot_Compare_Binomial_Poisson}
\alias{Plot_Compare_Binomial_Poisson}
\title{Draw a graph to compare the binomial and Possion distribution}
\usage{
Plot_Compare_Binomial_Poisson(pi = 0.1, size = 100)
}
\description{
Draw a graph to compare the binomial and Possion distribution
}
