% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Dummy_Relation.R
\name{Plot_Dummy_Relation}
\alias{Plot_Dummy_Relation}
\title{Plot possible relations between a factor and a quantative variable}
\usage{
Plot_Dummy_Relation(Adj_X = 0, Adj_Y = -100, Interaction = 1)
}
\arguments{
\item{Adj_X}{Difference on the X level}

\item{Adj_Y}{Difference on the Y level}

\item{Interaction}{Wheter the effects of the dummy variable and the quantative variable are interacted}
}
\description{
Illisturate different possible relations between two a factor and a quantative variable
}
\examples{
Plot_Dummy_Relation()

importFrom scales alpha
importFrom ggsci pal_lancet
importFrom nlme lmList

export
}
