% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_F_Distribution.R
\name{Plot_F_Distribution}
\alias{Plot_F_Distribution}
\title{Plot F Distribution}
\usage{
Plot_F_Distribution(
  df1 = 1,
  df2 = 14,
  alpha_level = 0.05,
  x_range = c(0, 7),
  y_range = NULL,
  fill_critical_region = TRUE,
  show_t_curve = FALSE
)
}
\description{
Plot F Distribution
}
