% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Factor_Relation.R
\name{Plot_Factor_Relation}
\alias{Plot_Factor_Relation}
\title{Plot possible relations between two factors}
\usage{
Plot_Factor_Relation(R = 2, C = 3, vls = 1:(R * C))
}
\arguments{
\item{R}{levels of the row factor}

\item{C}{levels of the column factor}

\item{vls}{values to be draw}
}
\description{
Illisturate different possible relations between two factors
}
\examples{
Plot_Factor_Relation()

}
