% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Geometric_Density.R
\name{Plot_Geometric_Density}
\alias{Plot_Geometric_Density}
\title{Plot the density of a geometric distribution}
\usage{
Plot_Geometric_Density(xx = 0:10, pi = c(0.3, 0.4, 0.5))
}
\description{
Plot the density of a geometric distribution
}
\examples{
Plot_Geometric_Density()
}
