% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Hypergeometric_Density.R
\name{Plot_Hypergeometric_Density}
\alias{Plot_Hypergeometric_Density}
\title{Plot the density of a hypergeometric distribution}
\usage{
Plot_Hypergeometric_Density(m = c(10, 40, 60), n = 10, k = c(5, 20),
  xx = 0:20, plot_density_function = TRUE)
}
\description{
Plot the density of a hypergeometric distribution
}
