% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Negative_Binomial_Density.R
\name{Plot_Negative_Binomial_Density}
\alias{Plot_Negative_Binomial_Density}
\title{Plot the density of a negative binomial distribution}
\usage{
Plot_Negative_Binomial_Density(
  r = c(1, 3, 5),
  pi = c(0.4, 0.6),
  plot_density_function = TRUE,
  xx = 0:20
)
}
\description{
Plot the density of a negative binomial distribution
}
