% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Poisson_Density.R
\name{Plot_Poisson_Density}
\alias{Plot_Poisson_Density}
\title{Plot the density of a Poisson distribution}
\usage{
Plot_Poisson_Density(xx = 0:20, lambda = c(4, 7, 10))
}
\description{
Plot the density of a Poisson distribution
}
