% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Population_Density.R
\name{Plot_Population_Density_Single}
\alias{Plot_Population_Density_Single}
\title{Plot density distributions of one population}
\usage{
Plot_Population_Density_Single(m, s, p = NULL, show_arrow = TRUE, ...)
}
\description{
Plot density distributions of one population
}
