% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Population_Density.R
\name{Plot_Population_Density_Single_Critical_Region}
\alias{Plot_Population_Density_Single_Critical_Region}
\title{Plot density distributions of one population with criteria}
\usage{
Plot_Population_Density_Single_Critical_Region(
  m = 0,
  s = 1,
  n = NULL,
  alpha = 0.05,
  two_tails = TRUE,
  p = 1.65
)
}
\description{
Plot density distributions of one population with criteria
}
