% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Quantile_Quantile_Population.R
\name{Plot_Population_Quantile_Quantile}
\alias{Plot_Population_Quantile_Quantile}
\title{Plot the theoretical quantile-quantile plot (q-q plot)}
\usage{
Plot_Population_Quantile_Quantile(family = c("c", "f", "g", "t", "n"),
  ..., adj = FALSE)
}
\arguments{
\item{family}{= c("c", "f", "g", "t", "n").}

\item{...}{arguments passed to the cdf function.}

\item{adj}{= FALSE whether to adjust the quantile of the function.}
}
\description{
Plot the theoretical quantile-quantile plot (q-q plot)
}
\examples{
Plot_Population_Quantile_Quantile("c", df = 3)
Plot_Population_Quantile_Quantile("t", df = 3)
Plot_Population_Quantile_Quantile("f", df1 = 5, df2 = 3, adj = TRUE)
}
