% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Quantile_Quantile_Sample.R
\name{Plot_Quantile_Quantile_Sample}
\alias{Plot_Quantile_Quantile_Sample}
\title{Plot the Empirical Quantile Quantile Diagram (q-q plot)}
\usage{
Plot_Quantile_Quantile_Sample(data)
}
\arguments{
\item{data}{Empirical Data returened by `ECDF`}
}
\description{
Plot the Empirical Quantile Quantile Diagram (q-q plot)
}
\examples{
set.seed(1)
norm100 <- rnorm(100, mean = 50, sd = 10)
qql(nrm100)
Plot_Quantile_Quantile_Sample(nrm100)
}
