% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_R_Source_Code.R
\name{Plot_R_Source_Code}
\alias{Plot_R_Source_Code}
\title{Plot the Number of lines in R source Code}
\usage{

  Plot_R_Source_Code(directory = "~/Documents/ADMIN/R-Source-Code/R-3.5.1/src")
}
\description{
Plot the Number of lines in R source Code
}
\examples{
Plot_R_Source_Code()
}
