% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_SoundWave_Graph.R
\name{Plot_Soundwave_Graph}
\alias{Plot_Soundwave_Graph}
\title{Plot the sound wave graph}
\usage{
Plot_Soundwave_Graph(
 
    Root_Directory = "~/Desktop/Conditional_New/Important_Information/Experiment_Structure"
)
}
\arguments{
\item{Root_Directory}{The directory that contains the sound files}
}
\description{
Draw the sound wave graphs in a directory
}
\examples{
Draw_Soundwave_Graph()
}
