% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_help_console.R
\name{print_help_console}
\alias{print_help_console}
\title{Print help information on console}
\usage{
print_help_console(topic, format = c("text", "html", "latex", "Rd"),
  lines = NULL, before = NULL, after = NULL)
}
\description{
Print help information on console
}
