\name{RegCod}
\alias{RegCod}
\title{Dummy and Deviation Coding of two factors in lm}
\description{
Dummy and Deviation Coding of two factors in lm
}
\arguments{
  \item{RowLevel}{Length of the Row factor}
  \item{ColLevel}{Length of the collumn factor}
  \item{Dummy}{if Dummy = TRUE, the default, it produces the dummy coding schema, otherwise, it produces the deviation schema}
  \item{RowNumber}{The number of observations in each level of the Row factor}
  \item{ColNumber}{The number of observations in each level of the Column factor}
}
\usage{
RegCod (RowLevel = 2, ColLevel = 3, Dummy = TRUE, RowNumber = 1, ColNumber = 1
}
\examples{
RegCod()
}
