% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Restrict_Overparameterized_Model.R
\name{Restrict_Overparameterized_Model}
\alias{Restrict_Overparameterized_Model}
\title{Restrict Over-parameterized model
Restrict the overparameterized model into appropriate one}
\usage{
Restrict_Overparameterized_Model(
  RowLevel = 2,
  ColLevel = 3,
  Dummy = TRUE,
  RowNumber = 1,
  ColNumber = 1
)
}
\description{
Restrict Over-parameterized model
Restrict the overparameterized model into appropriate one
}
