% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_All_Samples.R
\name{Sampling_Grid}
\alias{Sampling_Grid}
\title{List all samples}
\usage{
Sampling_Grid(data, size = NA, fun = NA, bp = 0.5)
}
\description{
List all samples
}
\examples{
x <- c(2, 4, 6, 8, 11, 15, 50, 60)
xx <- Sampling_Grid(x, size = 3)[["Sampled_Grid"]][["Statistic"]]
}
