\name{slptr}
\alias{slptr}
\title{Draw relations between two quantative variables and one category variable}
\description{
Draw relations between two quantative variables and one category variable
}
\arguments{
  \item{data}{data}
  \item{xlab, ylab, title}{xlab, ylab, Title of the diagram}
}
\usage{
slptr(
data = ideal1,
title = "(a). Gender and education are independent",
xlab = "Education",
ylab = "Income")
}
\examples{
slptr()
}
