% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_EngLish_Words.R
\name{Word3}
\alias{Word3}
\title{Add words to my dictionary}
\usage{

  Word3(Stored_Data_File = "~/Documents/ADMIN/Academic_Words/Word_Database.csv")
}
\description{
Add words to my dictionary
}
