% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeverageDiscrepancyInfluence.R
\name{LeverageDiscrepancyInfluence}
\alias{LeverageDiscrepancyInfluence}
\title{LeverageDiscrepancyInfluence}
\usage{
LeverageDiscrepancyInfluence(
  Points = c(mean(1:5), 1),
  n = 5,
  slope = 1,
  adj = 1
)
}
\description{
LeverageDiscrepancyInfluence
}
