% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Right_Eye_Data_Into_Left.R
\name{Merge_Right_Eye_Data_Into_Left}
\alias{Merge_Right_Eye_Data_Into_Left}
\title{Merge data obtained from Right eye into relevant columns of left columns}
\usage{
Merge_Right_Eye_Data_Into_Left(EDF)
}
\arguments{
\item{EDF}{Data outputed from Data Viewer samping report
Merge_Right_Eye_Data_Into_Left()}
}
\description{
Merge eye tracing data with eyelink
}
